+++ src/main/java/kr/nutee/nuteebackend/Controller/SearchController.java
@@ -1,19 +1,28 @@
 package kr.nutee.nuteebackend.Controller;
 
+import kr.nutee.nuteebackend.Service.PostService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping(path = "/sns/search",consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
+@RequestMapping(path = "/sns/search",consumes = MediaType.APPLICATION_JSON_VALUE)
 @RequiredArgsConstructor
 @ResponseBody
 @Slf4j
 public class SearchController {
 
-    @GetMapping(path = "/{text}")
-    public void searchPost(@PathVariable String text){
+    private final PostService postService;
 
+    @GetMapping(path = "/{text}")
+    public ResponseEntity<Object> searchPost(
+            @PathVariable String text,
+            @RequestParam("lastId") int lastId,
+            @RequestParam("limit") int limit
+    ){
+        return new ResponseEntity<>(postService.searchPost((long)lastId,limit,text), HttpStatus.OK);
     }
 }
